refactor: extract entity validation to separate module for line limit compliance
- Extract validateGitHubEntityExistence to src/github-entity-validation.lib.mjs
- Re-export from github.lib.mjs for backward compatibility
- Condense comments in solve.mjs and telegram-bot.mjs to stay under 1500 line limit
- All three files now at or under 1500 lines
